requirements.txt
streamlit==1.44.0
faiss-cpu==1.10.0
sentence-transformers==4.0.1
matplotlib==3.9.4
pandas==2.2.3
google-generativeai==0.8.4
python-dotenv==1.0.0
